+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support create an endpoint for listing fully hydrated NSGs attached to a given VLAN
 - Support ADBS: Customer Managed Keys (Integration Key Management System)
+- Support for OPSI External Databases added
 
 ## 4.28.0 (May 26, 2021)